refactor of packages and fix Readproperties class
