+++ src/main/java/org/fundacionjala/pivotal/util/WebDriverManager.java
@@ -1,6 +1,7 @@
-package org.fundacionjala.pivotal.model.pageobjects;
+package org.fundacionjala.pivotal.util;
 
 import org.fundacionjala.pivotal.ReadProperties;
+import org.fundacionjala.pivotal.browsers.BrowserFactory;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
@@ -27,7 +28,8 @@ public final class  WebDriverManager {
      * Constructor of class.
      */
     private WebDriverManager() {
-        this.driver = BrowserFactory.getBrowser(ReadProperties.getBrowser());
+        this.driver = BrowserFactory.getBrowser(ReadProperties.
+                getInstance().getProperties().getProperty("browser"));
         this.wait = new WebDriverWait(this.getDriver(), WAIT_TIME);
     }
